explosions for ship and asteroids
i am committing a scene with a few extra script in order to create the
explosions for both the asteroid as well as the ship i also messed
around with the background and stuff just to learn from the unity
tutorials. still isn't complete because it wasn't working some Vector3
error.